Bugfix/address broken 143 release (#16)
Remove dependency on `snmpy` and `build-pysnmp-mib`.

Unfortunately `build-pysnmp-mib` has proven to be quite unreliable.
In the last release (1.4.3), the output from that tool broke the
package since the format of the generated code changed
significantly. (The package has subsequently been pulled from
pypi.) This tool has also caused a number of issues in other
releases. So it's about time that that dependency disappear.

Bye-bye `build-pysnmp-mib`. You'll be missed.